added extension to users yaml file path. changed default from "dev" to "prod".
Both Marcus and Paul Miller have in the last 5 days had problems accessing instances to debug and develop.

Source of the issue is that the update_users cron script is failing when retrieving list of users from github repo due to missing file extension in Ansible template.

There is also an additional complexity that some instances do not provide the name of the bastion inventory (the var used in template) to the user_data script as an ansible var, hence it is falling back to a default of "dev.yml".

Will change that default to "prod".
Because we want prod to work correctly, dev instances with this issue can be fixed manually or the code updated so that the default is not required.
I would prefer to remove the default and identify where it is required but time is not available for that.